restructured files for organization, file names that have Classes should be capitalized according to Class name.
