+++ front-end/src/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
-import logo from './logo.svg';
 import './App.css';
-import Users from './components/users'
+import Users from './components/users/Users'
 import { Route, Link, Switch } from 'react-router-dom'
 
 class App extends Component {
