+++ front-end/src/components/users/NewUser.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import axios from 'axios'
-import "../../src/register-page.css"
+import "../../register-page.css"
 
 
 class NewUser extends React.Component {
